+++ cmd/agt/main.go
@@ -86,6 +86,8 @@ func run(args []string, stdout, stderr io.Writer) int {
 		return cmdStatus(args[1:], stdout, stderr)
 	case "warden":
 		return cmdWarden(args[1:], stdout, stderr)
+	case "redact":
+		return cmdRedact(args[1:], stdout, stderr)
 	case "doctor":
 		return cmdDoctor(args[1:], stdout, stderr)
 	case "quickstart":